impl simple simulation engine
Implement a simple simulation engine that is capable of running the
`four_pulses` test example. Modify the builder to produce a simualtion
state. Add a simulation engine that takes this state and drives its
progression through time. Tracers allow for the generated waveforms to
be captured and stored by some means.