improve event queue
Improve the event queue by moving from a binary heap to a BTreeMap from
time to values. The values are themselves a HashMap from signal pointers
to values. This makes it easy to extract all events associated with a
point in time at once, and also retains only the latest drive to a
signal in any given time slot.