improve entity scheduling
The builder now collapses all signals in instances into one vector, and
creates a reverse lookup table that associates signals with the
corresponding LLHD value. This is useful in order to trigger
instructions affected by a signal.

Improve the scheduling of entities. In the first step, trigger execution
of all instructions. For all subsequent steps, propagate signal changes
to the instructions using them. Also fix an issue in the value update
where a change in a signal would not propagate through a pointer
adjusting instruction.